fixup! fix: make core block-object behaviors container-aware
