+++ packages/editor/src/behaviors/behavior.abstract.keyboard.ts
@@ -1,9 +1,9 @@
 import {createKeyboardShortcut} from '@portabletext/keyboard-shortcuts'
 import {defaultKeyboardShortcuts} from '../editor/default-keyboard-shortcuts'
-import {getFocusTextBlock} from '../node-traversal/get-focus-text-block'
 import {getSibling} from '../node-traversal/get-sibling'
 import {getBlock} from '../node-traversal/is-block'
 import {getFocusInlineObject} from '../selectors/selector.get-focus-inline-object'
+import {getFocusTextBlock} from '../selectors/selector.get-focus-text-block'
 import {isSelectionCollapsed} from '../selectors/selector.is-selection-collapsed'
 import {isSelectionExpanded} from '../selectors/selector.is-selection-expanded'
 import {getBlockEndPoint} from '../utils/util.get-block-end-point'
