+++ packages/editor/src/behaviors/behavior.abstract.split.ts
@@ -1,8 +1,8 @@
 import {getAncestorTextBlock} from '../node-traversal/get-ancestor-text-block'
-import {getFocusBlockObject} from '../node-traversal/get-focus-block-object'
-import {getFocusTextBlock} from '../node-traversal/get-focus-text-block'
 import {isSelectionExpanded} from '../selectors'
+import {getFocusBlockObject} from '../selectors/selector.get-focus-block-object'
 import {getFocusInlineObject} from '../selectors/selector.get-focus-inline-object'
+import {getFocusTextBlock} from '../selectors/selector.get-focus-text-block'
 import {isTextBlockNode} from '../slate/node/is-text-block-node'
 import {isEqualSelectionPoints} from '../utils'
 import {parseBlock} from '../utils/parse-blocks'
