+++ packages/editor/src/behaviors/behavior.core.block-objects.ts
@@ -1,7 +1,7 @@
 import {defaultKeyboardShortcuts} from '../editor/default-keyboard-shortcuts'
-import {getFocusBlockObject} from '../node-traversal/get-focus-block-object'
-import {getFocusTextBlock} from '../node-traversal/get-focus-text-block'
 import {getSibling} from '../node-traversal/get-sibling'
+import {getFocusBlockObject} from '../selectors/selector.get-focus-block-object'
+import {getFocusTextBlock} from '../selectors/selector.get-focus-text-block'
 import {isSelectionCollapsed} from '../selectors/selector.is-selection-collapsed'
 import {isTextBlockNode} from '../slate/node/is-text-block-node'
 import {isListBlock} from '../utils/parse-blocks'
